chore: split CLI help data from its renderer
- What: the `BUILTIN_HELP` / `SUBCOMMAND_HELP` tables move to
  `src/cli/help-entries.ts`. `src/cli/help.ts` 294 → **86**, and its entry comes
  off the file-size allow-list rather than being lowered.

- Why: #2833. About 200 of those 294 lines were the tables. Data and the code that
  renders it are separate concerns — adding a command should not touch the
  renderer, and changing the layout should not touch the table.

- Impact: none. `help.ts` re-exports `BUILTIN_HELP` and `CliHelpEntry`, so existing
  importers are unaffected. `SUBCOMMAND_HELP` was file-private and is now exported
  from the new module, which is the only visibility change.

Verified the rendered output, not just the types — a subcommand entry is the case
that could quietly break, since it lived in the private table:

  builtinHelpFor('plugin install') -> FOUND
  "plugin install — install a CLI plugin"
  root help: 39 lines

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate tests/build/ tests/cli/
tests/bin/` 199 pass · `bun cli/src/cli.ts --help` renders.

Refs #2833
