+++ src/cli/help-entries.ts
@@ -1,0 +1,219 @@
+/**
+ * Help text for every built-in CLI command.
+ *
+ * Split out of `src/cli/help.ts` (#2833): that file was 294 lines against the 250-line rule,
+ * and ~200 of them were this table. Data and the code that renders it are genuinely separate
+ * concerns — adding a command should not touch the renderer, and changing the layout should not
+ * touch the table.
+ */
+export interface CliHelpEntry {
+  command: string;
+  help?: string;
+  aliases?: string[];
+  subcommands?: string[];
+  flags?: string[] | Record<string, string>;
+  usage?: string;
+  examples?: string[];
+}
+
+export const BUILTIN_HELP: CliHelpEntry[] = [
+  {
+    command: "plugin",
+    help: "manage installable CLI plugins",
+    usage: "arra-cli plugin <subcommand>",
+    subcommands: ["list", "info <name>", "install <url-or-path>", "remove <name>"],
+    flags: ["--json", "--yml", "--help", "-h"],
+    examples: ["arra-cli plugin list", "arra-cli plugin info search", "arra-cli plugin install ./my-plugin"],
+  },
+  {
+    command: "plugins",
+    help: "manage MCP tool plugin manifest",
+    usage: "arra-cli plugins <subcommand>",
+    subcommands: ["list", "enable <name>", "disable <name>"],
+    flags: ["--json", "--help", "-h"],
+    examples: ["arra-cli plugins list --json", "arra-cli plugins disable trace", "arra-cli plugins enable trace"],
+  },
+  {
+    command: "session",
+    help: "inspect sessions",
+    usage: "arra-cli session <subcommand>",
+    subcommands: ["list", "show <id>", "context <id>"],
+    flags: ["--json", "--yml", "--help", "-h"],
+    examples: ["arra-cli session list", "arra-cli session show abc123", "arra-cli session context abc123 --yml"],
+  },
+  {
+    command: "menu",
+    help: "inspect and customize studio menu",
+    usage: "arra-cli menu <subcommand>",
+    subcommands: ["list [--custom]", "add --path /p --label L [--group g] [--order N] [--icon i]", "remove <path>", "gist-status", "gist-url <url> [--override]", "gist-clear", "gist-reload", "reset-all [--yes]"],
+    flags: ["--json", "--yml", "--custom", "--override", "--yes", "--help", "-h"],
+    examples: ["arra-cli menu list --json", "arra-cli menu add --path /lab --label Lab", "arra-cli menu gist-url https://gist.github.com/user/id"],
+  },
+  {
+    command: "config",
+    help: "show resolved API target and config sources",
+    usage: "arra-cli config [show|path|use <name>]",
+    subcommands: ["show", "path", "use <name>"],
+    flags: ["--json", "--help", "-h"],
+    examples: ["arra-cli config show", "arra-cli config path", "arra-cli config use cafe"],
+  },
+  {
+    command: "doctor",
+    help: "verify backend, DB, FTS, vector, embedder, and MCP launch",
+    usage: "arra-cli doctor [--json]",
+    flags: ["--json", "--help", "-h"],
+    examples: ["arra-cli doctor", "arra-cli doctor --json", "arra-cli --at cafe doctor"],
+  },
+  {
+    command: "search",
+    help: "arra-cli search <query> [--type X] [--limit N] [--offset N] [--project P]",
+    usage: "arra-cli search <query> [--type X] [--limit N] [--offset N] [--project P]",
+    flags: ["--type", "--limit", "--offset", "--project", "--json", "--help", "-h"],
+    examples: [
+      "arra-cli search 'how does this work?'",
+      "arra-cli search oracle --type all --limit 20",
+    ],
+  },
+  {
+    command: "ask",
+    help: "ask the knowledge base about a question",
+    usage: "arra-cli ask <question> [--limit N] [--type X] [--model M] [--project P] [--as-of DATE] [--no-llm] [--json]",
+    flags: ["--limit", "--type", "--model", "--project", "--as-of", "--no-llm", "--json", "--help", "-h"],
+    examples: [
+      "arra-cli ask \"how does sleep mode work?\"",
+      "arra-cli ask \"what changed?\" --limit 5 --type pattern",
+    ],
+  },
+  {
+    command: "mcp",
+    help: "run MCP stdio server",
+    usage: "arra-cli mcp [--read-only]",
+    flags: ["--read-only", "--help", "-h"],
+    examples: ["arra-cli mcp", "arra-cli mcp --read-only"],
+  },
+  {
+    command: "install",
+    help: "install a plugin from URL or path",
+    usage: "arra-cli install <url-or-path> [flags]",
+    flags: ["--force", "--dry-run", "--artifact", "--manifest", "--no-backup", "--backup-dir", "--yml", "--help", "-h"],
+    examples: [
+      "arra-cli install ./my-plugin",
+      "arra-cli install github.com/owner/repo --dry-run",
+      "arra-cli install --artifact ./plugin.wasm --manifest ./plugin.json",
+    ],
+  },
+  {
+    command: "serve",
+    help: "start, stop, or inspect server status",
+    usage: "arra-cli serve [start|status|stop] [--foreground|--background] [--json]",
+    subcommands: ["start", "status", "stop"],
+    flags: ["--foreground", "-f", "--background", "-b", "--json", "--help", "-h"],
+    examples: [
+      "arra-cli serve",
+      "arra-cli serve status --json",
+      "arra-cli serve stop",
+    ],
+  },
+  {
+    command: "use",
+    help: "set the global default API target",
+    usage: "arra-cli use <name>",
+    examples: ["arra-cli use cafe"],
+  },
+  {
+    command: "completions",
+    help: "print shell completion scripts",
+    usage: "arra-cli completions <bash|zsh|fish>",
+    subcommands: ["bash", "zsh", "fish"],
+    flags: ["--help", "-h"],
+    examples: ["arra-cli completions zsh >> ~/.zshrc", "arra-cli completions fish"],
+  },
+  {
+    command: "mine",
+    help: "ingest a folder into Oracle memory",
+    usage: "arra-cli mine <dir> [--watch] [--db-path <file>] [--dry-run]",
+    flags: ["--watch", "--db-path <file>", "--dry-run", "--help", "-h"],
+    examples: ["arra-cli mine docs/sample-notes", "arra-cli mine ~/notes --watch"],
+  },
+  {
+    command: "huginn",
+    help: "run Huginn capture utilities",
+    usage: "arra-cli huginn <subcommand>",
+    subcommands: ["sweep"],
+    flags: ["--sessions-dir <path[:path...]>", "--repo-root <path>", "--lookback-hours <n>", "--max-files <n>", "--json", "--help", "-h"],
+    examples: ["arra-cli huginn sweep --lookback-hours 24", "arra-cli huginn sweep --sessions-dir ~/.claude/projects --json"],
+  },
+  {
+    command: "vector-config",
+    help: "inspect and manage vector embedding collection config",
+    usage: "arra-cli vector-config <subcommand>",
+    subcommands: ["list", "get <collection>", "stats [<collection>]", "set <collection> <field> <value>", "reload", "test <collection>"],
+    flags: ["--json", "--yml", "--help", "-h", "--model <name>", "--provider <name>", "--adapter <name>"],
+    examples: ["arra-cli vector-config list", "arra-cli vector-config get bge-m3", "arra-cli vector-config stats bge-m3", "arra-cli vector-config set bge-m3 enabled false", "arra-cli vector-config reload"],
+  },
+  {
+    command: "migrate",
+    help: "run Drizzle migration generate and push",
+    usage: "arra-cli migrate",
+    flags: ["--help", "-h"],
+    examples: ["arra-cli migrate"],
+  },
+  {
+    command: "seed",
+    help: "populate development DB sample data",
+    usage: "arra-cli seed",
+    flags: ["--help", "-h"],
+    examples: ["arra-cli seed"],
+  },
+  {
+    command: "backup",
+    help: "dump SQLite DB to timestamped SQL",
+    usage: "arra-cli backup [--out-dir <dir>]",
+    flags: ["--out-dir <dir>", "--help", "-h"],
+    examples: ["arra-cli backup", "arra-cli backup --out-dir ./backups"],
+  },
+  {
+    command: "changelog",
+    help: "generate CHANGELOG.md from git history",
+    usage: "arra-cli changelog [--since <tag>] [--out CHANGELOG.md] [--stdout]",
+    flags: ["--since <tag>", "--out <file>", "--stdout", "--help", "-h"],
+    examples: ["arra-cli changelog", "arra-cli changelog --since v1.2.3 --stdout"],
+  },
+  { command: "release", help: "bump CalVer, write changelog, and create a tag", usage: "arra-cli release [--beta|--stable] [--changelog CHANGELOG.md] [--dry-run]", flags: ["--beta", "--stable", "--changelog <file>", "--dry-run", "--check", "--help", "-h"], examples: ["arra-cli release", "arra-cli release --beta", "arra-cli release --dry-run"] },
+  { command: "export", help: "export a collection through Oracle v2", usage: "arra-cli export --url <url> --collection <name> --format markdown|json|jsonl --output <path>", flags: ["--url <url>", "--collection <name>", "--format <format>", "--output <path>", "--help", "-h"], examples: ["arra-cli export --url http://localhost:47778 --collection oracle_documents --format json --output oracle_documents.json"] },
+];
+
+export const SUBCOMMAND_HELP: CliHelpEntry[] = [
+  { command: "config show", help: "show resolved API target and config sources", usage: "arra-cli config show", examples: ["arra-cli config show"] },
+  { command: "config path", help: "print the global config path", usage: "arra-cli config path", examples: ["arra-cli config path"] },
+  { command: "config use", help: "set the global default API target", usage: "arra-cli config use <name>", examples: ["arra-cli config use cafe"] },
+  { command: "session list", help: "list all sessions", usage: "arra-cli session list [--json|--yml]", examples: ["arra-cli session list", "arra-cli session list --yml"] },
+  { command: "session show", help: "show a session summary", usage: "arra-cli session show <id> [--json|--yml]", examples: ["arra-cli session show abc123"] },
+  { command: "session context", help: "dump full session context", usage: "arra-cli session context <id> [--json|--yml]", examples: ["arra-cli session context abc123 --yml"] },
+  { command: "menu list", help: "list menu items", usage: "arra-cli menu list [--custom] [--json|--yml]", examples: ["arra-cli menu list", "arra-cli menu list --custom --json"] },
+  { command: "menu add", help: "add or replace a custom menu item", usage: "arra-cli menu add --path /p --label L [--group g] [--order N] [--icon i]", examples: ["arra-cli menu add --path /lab --label Lab --group tools"] },
+  { command: "menu remove", help: "remove a custom menu item", usage: "arra-cli menu remove <path>", examples: ["arra-cli menu remove /lab"] },
+  { command: "menu gist-status", help: "show current gist menu source", usage: "arra-cli menu gist-status", examples: ["arra-cli menu gist-status"] },
+  { command: "menu gist-url", help: "set gist menu URL", usage: "arra-cli menu gist-url <url> [--override]", examples: ["arra-cli menu gist-url https://gist.github.com/user/id"] },
+  { command: "menu gist-clear", help: "clear gist menu URL", usage: "arra-cli menu gist-clear", examples: ["arra-cli menu gist-clear"] },
+  { command: "menu gist-reload", help: "force refetch of gist menu", usage: "arra-cli menu gist-reload", examples: ["arra-cli menu gist-reload"] },
+  { command: "menu reset-all", help: "reset all custom menu state", usage: "arra-cli menu reset-all [--yes]", examples: ["arra-cli menu reset-all --yes"] },
+  { command: "vector-config list", help: "list available vector collections", usage: "arra-cli vector-config list", examples: ["arra-cli vector-config list"] },
+  { command: "vector-config get", help: "show a collection config", usage: "arra-cli vector-config get <collection>", examples: ["arra-cli vector-config get bge-m3"] },
+  { command: "vector-config stats", help: "show vector document counts", usage: "arra-cli vector-config stats [<collection>]", examples: ["arra-cli vector-config stats", "arra-cli vector-config stats bge-m3"] },
+  { command: "vector-config set", help: "set collection model/provider/adapter/enabled", usage: "arra-cli vector-config set <collection> <field> <value>", examples: ["arra-cli vector-config set bge-m3 enabled false", "arra-cli vector-config set bge-m3 --adapter qdrant --provider remote"] },
+  { command: "vector-config add", help: "add a vector collection config", usage: "arra-cli vector-config add <name> --model <name> [--collection <name>] [--adapter <name>]", examples: ["arra-cli vector-config add qwen4 --model qwen4-embedding --adapter lancedb"] },
+  { command: "vector-config remove", help: "remove a vector collection config", usage: "arra-cli vector-config remove <collection> [--yes]", examples: ["arra-cli vector-config remove qwen4 --yes"] },
+  { command: "vector-config set-primary", help: "mark a vector collection as primary", usage: "arra-cli vector-config set-primary <collection>", examples: ["arra-cli vector-config set-primary bge-m3"] },
+  { command: "vector-config reload", help: "clear cached vector collection stores", usage: "arra-cli vector-config reload", examples: ["arra-cli vector-config reload"] },
+  { command: "vector-config test", help: "probe one collection adapter", usage: "arra-cli vector-config test <collection>", examples: ["arra-cli vector-config test bge-m3"] },
+  { command: "plugins list", help: "list MCP tool plugins", usage: "arra-cli plugins list [--json]", examples: ["arra-cli plugins list --json"] },
+  { command: "plugins enable", help: "enable an MCP tool plugin", usage: "arra-cli plugins enable <name> [--json]", examples: ["arra-cli plugins enable trace"] },
+  { command: "plugins disable", help: "disable an MCP tool plugin", usage: "arra-cli plugins disable <name> [--json]", examples: ["arra-cli plugins disable trace"] },
+  { command: "plugin list", help: "list installed CLI plugins", usage: "arra-cli plugin list [--json|--yml]", examples: ["arra-cli plugin list"] },
+  { command: "plugin info", help: "show installed plugin details", usage: "arra-cli plugin info <name> [--json|--yml]", examples: ["arra-cli plugin info search"] },
+  { command: "plugin install", help: "install a CLI plugin", usage: "arra-cli plugin install <url-or-path> [--force] [--dry-run]", examples: ["arra-cli plugin install ./my-plugin --dry-run", "arra-cli plugin install github.com/org/plugin --force"] },
+  { command: "plugin remove", help: "remove an installed CLI plugin", usage: "arra-cli plugin remove <name> [--yes]", examples: ["arra-cli plugin remove search --yes"] },
+  { command: "huginn sweep", help: "back-fill missed Huginn session captures and learnings", usage: "arra-cli huginn sweep [--sessions-dir PATH[:PATH...]] [--repo-root PATH] [--lookback-hours N] [--max-files N] [--json]", examples: ["arra-cli huginn sweep --lookback-hours 24"] },
+];
+
